+++ CLI/src/main/java/com/example/CLI/Parser/ParserBuilder.java
@@ -26,11 +26,7 @@ public class ParserBuilder {
         var rules = new ArrayList<Rule>();
         rules.add(new PipeRule(informant));
         rules.add(new SaveRule(context));
-        rules.add(new EchoRule());
-        rules.add(new CatRule(informant));
-        rules.add(new PwdRule());
-        rules.add(new WcRule(informant));
-        rules.add(new UndefinedRule(context, informant));
+        rules.add(new CommandRule(context, informant));
         rules.add(new ExtractRule(context));
         rules.add(new DoubleQuotesRule(context));
         rules.add(new SingleQuotesRule());
