+++ CLI/src/main/java/com/example/CLI/Parser/Rules/CommandRule.java
@@ -1,13 +1,19 @@
 package com.example.CLI.Parser.Rules;
 
+import com.example.CLI.Commands.Operation;
+import com.example.CLI.Commands.Undefined;
+import com.example.CLI.Environment.Context;
+import com.example.CLI.Environment.Informant;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.regex.Pattern;
 
 
-public abstract class CommandRule implements Rule {
+public class CommandRule implements Rule {
 
+    @NotNull private Context context;
+    @NotNull private Informant informant;
     protected static final String regex =
             "\\s*([\\w./+=()*~$\\-]+|'[^']+'|\"(\\\\\"|[^\\\\\"])+\")" +
             "(\\s+([\\w./+=()*~$\\-]+|'[^']+'|\"(\\\\\"|[^\\\\\"])+\"))*\\s*";
@@ -17,9 +23,25 @@ public abstract class CommandRule implements Rule {
         word = Pattern.compile("[\\w./+=()*~$\\-]+|'[^']+'|\"(\\\\\"|[^\\\\\"])+\"");
     }
 
+    public CommandRule(@NotNull Context context, @NotNull Informant informant) {
+        this.context = context;
+        this.informant = informant;
+    }
+
     @Override @NotNull
     public Type getType() {
         return Type.COMMAND;
+    }
+
+    @NotNull
+    @Override
+    public Integer getLevel() {
+        return 10;
+    }
+
+    @Override
+    public boolean isMatching(@NotNull String input) {
+        return input.matches(regex);
     }
 
     @Override @NotNull
@@ -35,5 +57,15 @@ public abstract class CommandRule implements Rule {
         }
 
         return result;
+    }
+
+    @NotNull
+    @Override
+    public Operation createOperation(ArrayList<Operation> args) {
+        var command = new Undefined(context, args.get(0), informant);
+        args.remove(0);
+        command.setArgs(args);
+
+        return command;
     }
 }
